\name{weaverWSLOptimiser-package}
\alias{weaverWSLOptimiser-package}
\alias{weaverWSLOptimiser}
\docType{package}
\title{
\packageTitle{weaverWSLOptimiser}
}
\description{
\packageDescription{weaverWSLOptimiser}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{weaverWSLOptimiser}
\packageIndices{weaverWSLOptimiser}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{weaverWSLOptimiser}

Maintainer: \packageMaintainer{weaverWSLOptimiser}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
